SSN in Appointment changed to long so it can carry the actual ssn

+++ src/shared/Appointment.java
@@ -7,10 +7,10 @@ public class Appointment implements Serializable
 {
   private Timestamp from;
   private Timestamp to;
-  private int doctorSSN;
-  private int patientSSN;
+  private long doctorSSN;
+  private long patientSSN;
 
-  public Appointment(Timestamp from, Timestamp to, int doctorSSN, int patientSSN)
+  public Appointment(Timestamp from, Timestamp to, long doctorSSN, long patientSSN)
   {
     this.from = from;
     this.to = to;
@@ -28,12 +28,12 @@ public class Appointment implements Serializable
     return to;
   }
 
-  public int getDoctorSSN()
+  public long getDoctorSSN()
   {
     return doctorSSN;
   }
 
-  public int getPatientSSN()
+  public long getPatientSSN()
   {
     return patientSSN;
   }
@@ -48,12 +48,12 @@ public class Appointment implements Serializable
     this.to = to;
   }
 
-  public void setDoctorSSN(int doctorSSN)
+  public void setDoctorSSN(long doctorSSN)
   {
     this.doctorSSN = doctorSSN;
   }
 
-  public void setPatientSSN(int patientSSN)
+  public void setPatientSSN(long patientSSN)
   {
     this.patientSSN = patientSSN;
   }
